Move dev bundle ID patch to prestart to fix empty window on first npm start
