+++ apps/studio/package.json
@@ -12,7 +12,7 @@
 	},
 	"scripts": {
 		"start": "electron-vite dev --config ./electron.vite.config.ts --outDir=dist --watch",
-		"prestart": "npm -w wp-studio run build",
+		"prestart": "npm -w wp-studio run build && node ../../scripts/ensure-dev-bundle-id.mjs",
 		"make": "electron-vite build --config ./electron.vite.config.ts --outDir=dist && electron-forge make .",
 		"make:windows-x64": "electron-vite build --config ./electron.vite.config.ts --outDir=dist && electron-forge make . --arch=x64 --platform=win32",
 		"make:windows-arm64": "electron-vite build --config ./electron.vite.config.ts --outDir=dist && electron-forge make . --arch=arm64 --platform=win32",
